bot-reserva/requirements.txt
botbuilder-integration-aiohttp>=4.14.0
botbuilder-dialogs>=4.15.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiounittest==1.3.0
attrs==25.3.0
azure-core==1.35.1
Babel==2.9.1
botbuilder-core==4.17.0
botbuilder-dialogs==4.17.0
botbuilder-integration-aiohttp==4.17.0
botbuilder-schema==4.17.0
botframework-connector==4.17.0
botframework-streaming==4.17.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
datedelta==1.4
emoji==1.7.0
frozenlist==1.7.0
grapheme==0.6.0
idna==3.10
isodate==0.7.2
jsonpickle==1.4.2
msal==1.34.0
msrest==0.7.1
multidict==6.6.4
multipledispatch==1.0.0
oauthlib==3.3.1
propcache==0.4.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
pytz==2025.2
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.21.0

